requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pymupdf>=1.24.7
tqdm>=4.66.5

chromadb>=0.5.5
numpy>=1.26.4

rank-bm25>=0.2.2
streamlit>=1.36.0


